move dimension reduction func to representation
