+++ .github/workflows/build-and-release.yaml
@@ -20,10 +20,11 @@ jobs:
       run: |
         make init
         make install
-    - name: Build package & publish
+    - name: Build wheel
       run: |
         make wheel
-        make release
-      env:
-        user: sorrychoe
-        password: ${{ secrets.PYPI_PASSWORD }}
+    - name: Publish to Pypi
+      run: |
+          python3 -m pip install twine
+
+          python3 -m twine upload dist/*.whl --non-interactive -u ${{ secrets.PYPI_USERNAME }} -p ${{ secrets.PYPI_PASSWORD }}
